+++ app/src/main/java/com/wecast/mobile/ui/screen/vod/player/VodPlayerActivity.java
@@ -69,7 +69,7 @@ public class VodPlayerActivity extends BaseActivity<ActivityVodPlayerBinding, Vo
     private DebugTextViewHelper debugViewHelper;
     private WeExoPlayer weExoPlayer;
     private long bufferTime = 0;
-    private int seekTo = -1;
+    private float seekTo = -1;
     // Playlist continuity
     private VodPlayerNextEpisodeView nextEpisodeView;
     private ImageButton nextEpisode;
@@ -78,7 +78,7 @@ public class VodPlayerActivity extends BaseActivity<ActivityVodPlayerBinding, Vo
     private Runnable nextEpisodeRunnable = () -> nextEpisodeView.startCounter();
     private List<Vod> episodes;
 
-    public static void open(Context context, Vod item, VodSourceProfile profile, int playAction, int seekTo) {
+    public static void open(Context context, Vod item, VodSourceProfile profile, int playAction, float seekTo) {
         Intent intent = new Intent(context, VodPlayerActivity.class);
         intent.putExtra("ID", item.getId());
         intent.putExtra("IS_EPISODE", item.getMultiEventVodId() != 0);
@@ -129,7 +129,7 @@ public class VodPlayerActivity extends BaseActivity<ActivityVodPlayerBinding, Vo
             profileId = bundle.getInt("PROFILE_ID");
             profileBusinessModel = bundle.getString("BUSINESS_MODEL");
             playAction = bundle.getInt("PLAY_ACTION");
-            seekTo = bundle.getInt("SEEK_TO");
+            seekTo = bundle.getFloat("SEEK_TO");
             getById(id, isEpisode);
         }
 
@@ -583,7 +583,7 @@ public class VodPlayerActivity extends BaseActivity<ActivityVodPlayerBinding, Vo
      */
     private void seekToPosition() {
         if (seekTo > -1) {
-            weExoPlayer.seekToPosition(seekTo);
+            weExoPlayer.seekToPosition((int) seekTo);
         }
     }
